backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-adk-agents==1.0.0
python-multipart==0.0.6
pydantic==2.5.0